Unify aggregator benchmark code in aggregator_test
