+++ sdk/metric/internal/aggregator_test.go
@@ -18,6 +18,7 @@
 package internal // import "go.opentelemetry.io/otel/sdk/metric/internal"
 
 import (
+	"strconv"
 	"sync"
 	"testing"
 
@@ -128,4 +129,55 @@ func assertSetMap[N int64 | float64](t *testing.T, expected, actual setMap[N]) {
 	}
 
 	assert.Lenf(t, extra, 0, "unknown values added: %v", extra)
+}
+
+var bmarkResults []Aggregation
+
+func benchmarkAggregatorN[N int64 | float64](b *testing.B, factory func() Aggregator[N], count int) {
+	attrs := make([]attribute.Set, count)
+	for i := range attrs {
+		attrs[i] = attribute.NewSet(attribute.Int("value", i))
+	}
+
+	b.Run("Aggregate", func(b *testing.B) {
+		agg := factory()
+		b.ReportAllocs()
+		b.ResetTimer()
+
+		for n := 0; n < b.N; n++ {
+			for _, attr := range attrs {
+				agg.Aggregate(1, attr)
+			}
+		}
+		assert.Len(b, agg.Aggregations(), count)
+	})
+
+	b.Run("Aggregations", func(b *testing.B) {
+		aggs := make([]Aggregator[N], b.N)
+		for n := range aggs {
+			a := factory()
+			for _, attr := range attrs {
+				a.Aggregate(1, attr)
+			}
+			aggs[n] = a
+		}
+
+		b.ReportAllocs()
+		b.ResetTimer()
+
+		for n := 0; n < b.N; n++ {
+			bmarkResults = aggs[n].Aggregations()
+		}
+	})
+}
+
+func benchmarkAggregator[N int64 | float64](factory func() Aggregator[N]) func(*testing.B) {
+	counts := []int{1, 10, 100}
+	return func(b *testing.B) {
+		for _, n := range counts {
+			b.Run(strconv.Itoa(n), func(b *testing.B) {
+				benchmarkAggregatorN(b, factory, n)
+			})
+		}
+	}
 }
