+++ sdk/metric/internal/sum_test.go
@@ -18,12 +18,7 @@
 package internal // import "go.opentelemetry.io/otel/sdk/metric/internal"
 
 import (
-	"strconv"
 	"testing"
-
-	"github.com/stretchr/testify/assert"
-
-	"go.opentelemetry.io/otel/attribute"
 )
 
 func TestSum(t *testing.T) {
@@ -83,57 +78,6 @@ func testDeltaSumReset[N int64 | float64](a Aggregator[N]) func(*testing.T) {
 func TestDeltaSumReset(t *testing.T) {
 	t.Run("Int64", testDeltaSumReset(NewDeltaSum[int64]()))
 	t.Run("Float64", testDeltaSumReset(NewDeltaSum[float64]()))
-}
-
-var result []Aggregation
-
-func benchmarkAggregatorN[N int64 | float64](b *testing.B, factory func() Aggregator[N], count int) {
-	attrs := make([]attribute.Set, count)
-	for i := range attrs {
-		attrs[i] = attribute.NewSet(attribute.Int("value", i))
-	}
-
-	b.Run("Aggregate", func(b *testing.B) {
-		agg := factory()
-		b.ReportAllocs()
-		b.ResetTimer()
-
-		for n := 0; n < b.N; n++ {
-			for _, attr := range attrs {
-				agg.Aggregate(1, attr)
-			}
-		}
-		assert.Len(b, agg.Aggregations(), count)
-	})
-
-	b.Run("Aggregations", func(b *testing.B) {
-		aggs := make([]Aggregator[N], b.N)
-		for n := range aggs {
-			a := factory()
-			for _, attr := range attrs {
-				a.Aggregate(1, attr)
-			}
-			aggs[n] = a
-		}
-
-		b.ReportAllocs()
-		b.ResetTimer()
-
-		for n := 0; n < b.N; n++ {
-			result = aggs[n].Aggregations()
-		}
-	})
-}
-
-func benchmarkAggregator[N int64 | float64](factory func() Aggregator[N]) func(*testing.B) {
-	counts := []int{1, 10, 100}
-	return func(b *testing.B) {
-		for _, n := range counts {
-			b.Run(strconv.Itoa(n), func(b *testing.B) {
-				benchmarkAggregatorN(b, factory, n)
-			})
-		}
-	}
 }
 
 func BenchmarkSum(b *testing.B) {
